requirements.txt
Django==5.0.11
djangorestframework==3.15.2
psycopg2==2.9.9
django-jazzmin==3.0.1
drf-spectacular==0.28.0
Pillow==11.1.0